+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Support for the old dispatching: `#[teloxide(subtransition)]`.
 
+### Deprecated
+
+ - `#[derive(DialogueState)]` in favour of `teloxide::handler!`.
+
 ## 0.5.1 - 2022-03-23
 
 ### Fixed