+++ CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Rename `BotCommand` trait to `BotCommands` [**BC**].
  - `BotCommands::descriptions` now returns `CommandDescriptions` instead of `String` [**BC**].
 
+### Deprecated
+
+ - `HandlerFactory` and `HandlerExt::dispatch_by` in favour of `teloxide::handler!`.
+
 ## 0.7.2 - 2022-03-23
 
 ### Added