+++ CHANGELOG.md
@@ -13,8 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `User::{tme_url, preferably_tme_url}` convenience functions ([#197][pr197])
 - `Me::username` and `Deref<Target = User>` implementation for `Me` ([#197][pr197])
 - `Me::{mention, tme_url}` ([#197][pr197])
+- `AllowedUpdate::ChatJoinRequest` ([#201][pr201])
 
 [pr197]: https://github.com/teloxide/teloxide-core/pull/197
+[pr201]: https://github.com/teloxide/teloxide-core/pull/201
 
 ### Changed